Update version1.md by getting rid of the 2nd level header
Update the file called "version1.md" by getting rid of the 2nd level header: considering there was two headers, one with "Version 1.0" and the second one with "Version V1.0 of COMOKIT is now available"; considering there was no text between the 2; considering that the message needed to be as short and as straightforward as possible, I sat down and thought for some time. Then I took the (risky) decision to merge the two headers into one.  The result is entirely contained in this commit.